requirements.txt
langchain
langchain-community
langchain-google-genai
faiss-cpu
sentence-transformers
PyPDF2
python-docx
python-dotenv
numpy
pypdf
